+++ engine/src/semulator/core/v3/SEmulatorEngineV3.java
@@ -13,11 +13,11 @@ public interface SEmulatorEngineV3 {
     List<ProgramInfo> getPrograms();
     List<FunctionInfo> getFunctions();
     ProgramFunctionDto expand(String programName, boolean isProgramBool, int degreeOfExpand);
-    ExecutionRunDto runProgram(int credits, String username, String architecture, int desiredDegreeOfExpand, String programName, boolean isProgramBool, Map<String, Long> originalInputs, long ... input);
+    ExecutionRunDto runProgram(long credits, String username, String architecture, int desiredDegreeOfExpand, String programName, boolean isProgramBool, Map<String, Long> originalInputs, long ... input);
     DebugContextDto initialStartOfDebugger(String username, String programName, boolean isProgram, int degreeOfRun, DebugContextDto debugContext, Map<String, Long> originalInputs, long ... inputs);
-    DebugContextDto debug (int credits, String username, String programName, boolean isProgram, int degreeOfExpand, DebugContextDto debugContext, Map<String, Long> originalInputs);
+    DebugContextDto debug (long credits, String username, String programName, boolean isProgram, int degreeOfExpand, DebugContextDto debugContext, Map<String, Long> originalInputs);
     void addCurrentRunToHistory(DebugContextDto debugContext, int degreeOfRun, String programName, boolean isProgram, String architecture);
     List<RunResultDto> getUserRunHistory(String user);
-    DebugContextDto resume (int credits, String username, String programName, boolean isProgram, int degreeOfExpand, DebugContextDto debugContext, Map<String, Long> originalInputs);
+    DebugContextDto resume (long credits, String username, String programName, boolean isProgram, int degreeOfExpand, DebugContextDto debugContext, Map<String, Long> originalInputs);
 
 }
