+++ engine/src/semulator/logic/debugger/ProgramDebugger.java
@@ -5,8 +5,8 @@ import dto.DebugContextDto;
 import java.util.Map;
 
 public interface ProgramDebugger {
-    DebugContextDto debug (int credits, long instructionToExecuteNumber, DebugContextDto debugDetails, Map<String, Long> originalInputs);
-    DebugContextDto resume (int credits, long instructionToExecuteNumber, DebugContextDto debugDetails, Map<String, Long> originalInputs);
+    DebugContextDto debug (long credits, long instructionToExecuteNumber, DebugContextDto debugDetails, Map<String, Long> originalInputs);
+    DebugContextDto resume (long credits, long instructionToExecuteNumber, DebugContextDto debugDetails, Map<String, Long> originalInputs);
     DebugContextDto initialDebugger(Map<String, Long> originalInputs);
     DebugContextDto breakPointMode(long breakPointIndex, DebugContextDto result, Map<String, Long> originalInputs);
 }
